+++ crates/environ/src/component/artifacts.rs
@@ -2,7 +2,7 @@
 //! which are serialized with `bincode` into output ELF files.
 
 use crate::{
-    component::{AllCallFunc, Component, ComponentTypes, TrampolineIndex, TypeComponentIndex},
+    component::{Component, ComponentTypes, TrampolineIndex, TypeComponentIndex},
     CompiledModuleInfo, FunctionLoc, PrimaryMap, StaticModuleIndex,
 };
 use serde_derive::{Deserialize, Serialize};
@@ -42,4 +42,31 @@ pub struct CompiledComponentInfo {
     /// The location of the wasm-to-native trampoline for the `resource.drop`
     /// intrinsic.
     pub resource_drop_wasm_to_native_trampoline: Option<FunctionLoc>,
+}
+
+/// A triple of related functions/trampolines variants with differing calling
+/// conventions: `{wasm,array,native}_call`.
+///
+/// Generic so we can use this with either the `Box<dyn Any + Send>`s that
+/// implementations of the compiler trait return or with `FunctionLoc`s inside
+/// an object file, for example.
+#[derive(Clone, Copy, Default, Serialize, Deserialize)]
+pub struct AllCallFunc<T> {
+    /// The function exposing the Wasm calling convention.
+    pub wasm_call: T,
+    /// The function exposing the array calling convention.
+    pub array_call: T,
+    /// The function exposing the native calling convention.
+    pub native_call: T,
+}
+
+impl<T> AllCallFunc<T> {
+    /// Map an `AllCallFunc<T>` into an `AllCallFunc<U>`.
+    pub fn map<U>(self, mut f: impl FnMut(T) -> U) -> AllCallFunc<U> {
+        AllCallFunc {
+            wasm_call: f(self.wasm_call),
+            array_call: f(self.array_call),
+            native_call: f(self.native_call),
+        }
+    }
 }
