+++ crates/wasmtime/src/compile/code_builder.rs
@@ -1,15 +1,7 @@
-#[cfg(all(feature = "runtime", feature = "component-model"))]
-use crate::component::Component;
 use crate::Engine;
-#[cfg(feature = "runtime")]
-use crate::{instantiate::MmapVecWrapper, CodeMemory, Module};
 use anyhow::{anyhow, bail, Context, Result};
 use std::borrow::Cow;
 use std::path::Path;
-use std::sync::Arc;
-use wasmtime_environ::ObjectKind;
-#[cfg(feature = "runtime")]
-use wasmtime_runtime::MmapVec;
 
 /// Builder-style structure used to create a [`Module`](crate::Module) or
 /// pre-compile a module to a serialized list of bytes.
@@ -45,7 +37,7 @@ use wasmtime_runtime::MmapVec;
 /// [`wasm`]: CodeBuilder::wasm
 /// [`wasm_file`]: CodeBuilder::wasm_file
 pub struct CodeBuilder<'a> {
-    engine: &'a Engine,
+    pub(super) engine: &'a Engine,
     wasm: Option<Cow<'a, [u8]>>,
     wasm_path: Option<Cow<'a, Path>>,
     wat: bool,
@@ -130,7 +122,7 @@ impl<'a> CodeBuilder<'a> {
         Ok(self)
     }
 
-    fn wasm_binary(&self) -> Result<Cow<'_, [u8]>> {
+    pub(super) fn wasm_binary(&self) -> Result<Cow<'_, [u8]>> {
         let wasm = self
             .wasm
             .as_ref()
@@ -145,65 +137,6 @@ impl<'a> CodeBuilder<'a> {
             });
         }
         Ok((&wasm[..]).into())
-    }
-
-    #[cfg(feature = "runtime")]
-    fn compile_cached<T>(
-        &self,
-        build_artifacts: fn(&Engine, &[u8]) -> Result<(MmapVecWrapper, Option<T>)>,
-    ) -> Result<(Arc<CodeMemory>, Option<T>)> {
-        let wasm = self.wasm_binary()?;
-
-        self.engine
-            .check_compatible_with_native_host()
-            .context("compilation settings are not compatible with the native host")?;
-
-        #[cfg(feature = "cache")]
-        {
-            let state = (
-                HashedEngineCompileEnv(self.engine),
-                &wasm,
-                // Don't hash this as it's just its own "pure" function pointer.
-                NotHashed(build_artifacts),
-            );
-            let (code, info_and_types) =
-                wasmtime_cache::ModuleCacheEntry::new("wasmtime", self.engine.cache_config())
-                    .get_data_raw(
-                        &state,
-                        // Cache miss, compute the actual artifacts
-                        |(engine, wasm, build_artifacts)| -> Result<_> {
-                            let (mmap, info) = (build_artifacts.0)(engine.0, wasm)?;
-                            let code = publish_mmap(mmap.0)?;
-                            Ok((code, info))
-                        },
-                        // Implementation of how to serialize artifacts
-                        |(_engine, _wasm, _), (code, _info_and_types)| Some(code.mmap().to_vec()),
-                        // Cache hit, deserialize the provided artifacts
-                        |(engine, wasm, _), serialized_bytes| {
-                            let kind = if wasmparser::Parser::is_component(&wasm) {
-                                ObjectKind::Component
-                            } else {
-                                ObjectKind::Module
-                            };
-                            let code = engine.0.load_code_bytes(&serialized_bytes, kind).ok()?;
-                            Some((code, None))
-                        },
-                    )?;
-            return Ok((code, info_and_types));
-        }
-
-        #[cfg(not(feature = "cache"))]
-        {
-            let (mmap, info_and_types) = build_artifacts(self.engine, &wasm)?;
-            let code = publish_mmap(mmap.0)?;
-            return Ok((code, info_and_types));
-        }
-
-        struct NotHashed<T>(T);
-
-        impl<T> std::hash::Hash for NotHashed<T> {
-            fn hash<H: std::hash::Hasher>(&self, _hasher: &mut H) {}
-        }
     }
 
     /// Finishes this compilation and produces a serialized list of bytes.
@@ -228,18 +161,6 @@ impl<'a> CodeBuilder<'a> {
         Ok(v)
     }
 
-    /// Same as [`CodeBuilder::compile_module_serialized`] except that a
-    /// [`Module`](crate::Module) is produced instead.
-    ///
-    /// Note that this method will cache compilations if the `cache` feature is
-    /// enabled and turned on in [`Config`](crate::Config).
-    #[cfg(feature = "runtime")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "runtime")))]
-    pub fn compile_module(&self) -> Result<Module> {
-        let (code, info_and_types) = self.compile_cached(super::build_artifacts)?;
-        Module::from_parts(self.engine, code, info_and_types)
-    }
-
     /// Same as [`CodeBuilder::compile_module_serialized`] except that it
     /// compiles a serialized [`Component`] instead of a module.
     #[cfg(feature = "component-model")]
@@ -248,18 +169,6 @@ impl<'a> CodeBuilder<'a> {
         let bytes = self.wasm_binary()?;
         let (v, _) = super::build_component_artifacts(self.engine, &bytes)?;
         Ok(v)
-    }
-
-    /// Same as [`CodeBuilder::compile_module`] except that it compiles a
-    /// [`Component`] instead of a module.
-    #[cfg(all(feature = "runtime", feature = "component-model"))]
-    #[cfg_attr(
-        docsrs,
-        doc(cfg(all(feature = "runtime", feature = "component-model")))
-    )]
-    pub fn compile_component(&self) -> Result<Component> {
-        let (code, artifacts) = self.compile_cached(super::build_component_artifacts)?;
-        Component::from_parts(self.engine, code, artifacts)
     }
 }
 
@@ -288,11 +197,4 @@ impl std::hash::Hash for HashedEngineCompileEnv<'_> {
         // Catch accidental bugs of reusing across crate versions.
         config.module_version.hash(hasher);
     }
-}
-
-#[cfg(feature = "runtime")]
-fn publish_mmap(mmap: MmapVec) -> Result<Arc<CodeMemory>> {
-    let mut code = CodeMemory::new(mmap)?;
-    code.publish()?;
-    Ok(Arc::new(code))
 }
